+++ packages/react/src/atomic/Input/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The Input component allows the user to enter or edit information in text format.
 
+## 2025-08-20 `2.6.0`
+
+#### 🎉 New features
+
+- Added new `ai-generative` appearance to `Input`, `Input.Search` and `Input.Password` to match Nimbus AI gradient style. ([#TBD](https://github.com/TiendaNube/nimbus-design-system/pull/TBD) by [@joacotornello](https://github.com/joacotornello))
+
 ## 2025-03-18 `2.5.0`
 
 #### 🎉 New features